Add GET '/requests' endpoint, getRequestsByUser service & tests

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/templatepackagename/model/LeaveRequest.kt
@@ -14,7 +14,7 @@ data class LeaveRequest(
 
   val startDate: LocalDate,
   val endDate: LocalDate,
-  val duration: Int,
+  val duration: Double,
   val isFirstDayHalfDay: Boolean,
   val isLastDayHalfDay: Boolean,
